requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.1.0
langchain-ollama>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pymupdf>=1.24.0
openai-whisper>=20231117
pytesseract>=0.3.10
pillow>=10.0.0
pydantic>=2.7.0
python-multipart>=0.0.9
httpx>=0.27.0
python-dotenv>=1.0.0
langsmith>=0.1.0
datasets>=2.19.0
pytest>=8.2.0
torch>=2.2.0
pandas>=2.2.0
scikit-learn>=1.4.0
pydicom>=2.4.4
